+++ src/views/ClientParticleFunctions.jsx
@@ -9,15 +9,14 @@ const { getClientDevices } = require('../utilities/DatabaseFunctions')
 const { callClientParticleFunction, getFirmwareVersion, getFunctionList } = require('../utilities/ParticleFunctions')
 
 const styles = {
-  column: {
-    flex: '0 0 35%',
-    padding: 20,
-    alignItems: 'top',
+  pageContainer: {
     display: 'flex',
     flexDirection: 'column',
+    height: '100vh',
+    overflowY: 'auto',
+    padding: '20px',
   },
-  scrollView: {
-    overflow: 'auto',
+  contentContainer: {
     paddingRight: '5px',
     paddingLeft: '5px',
     paddingBottom: '5px',
@@ -30,7 +29,7 @@ const styles = {
     padding: '2px 0',
   },
   selectAll: {
-    margin: '8px 0px 8px 0px',
+    margin: '8px 0',
   },
 }
 
@@ -85,8 +84,9 @@ function ClientParticleFunctions(props) {
     event.preventDefault()
 
     setLoadingFetch(true)
-    setClientData({ functionName: '', argument: '' })
     setDevices({ all: [], selected: [] })
+    setClientData({ functionName: '', argument: '' })
+    setFunctionList([])
     setAlerts([])
 
     try {
@@ -112,6 +112,7 @@ function ClientParticleFunctions(props) {
 
     setLoadingFunctions(true)
     setClientData({ functionName: '', argument: '' })
+    setFunctionList([])
     setAlerts([])
 
     try {
@@ -215,7 +216,7 @@ function ClientParticleFunctions(props) {
   }
 
   return (
-    <div style={styles.column}>
+    <div style={styles.pageContainer}>
       <div>
         <h3>Client Particle Functions</h3>
         <hr />
@@ -237,7 +238,7 @@ function ClientParticleFunctions(props) {
         </Alert>
       ))}
 
-      <div style={styles.scrollView}>
+      <div style={styles.contentContainer}>
         {/* fetch devices from database */}
         <Form onSubmit={handleFetchDevices}>
           <Form.Group className="mb-3" controlId="formDisplayName">